# Swagger/swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"gift-certificate-controller","description":"Gift Certificate Controller"},{"name":"order-controller","description":"Order Controller"},{"name":"tag-controller","description":"Tag Controller"},{"name":"user-controller","description":"User Controller"}],"paths":{"/certificate":{"get":{"tags":["gift-certificate-controller"],"summary":"showAll","operationId":"showAllUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["gift-certificate-controller"],"summary":"createCertificate","operationId":"createCertificateUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"giftCertificate","description":"giftCertificate","required":true,"schema":{"$ref":"#/definitions/GiftCertificate"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/certificate/search/by-part-name-description":{"get":{"tags":["gift-certificate-controller"],"summary":"getCertificatesByPartOfNameOrDescription","operationId":"getCertificatesByPartOfNameOrDescriptionUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"part","in":"query","description":"part","required":true,"type":"string"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/certificate/search/by-tag":{"get":{"tags":["gift-certificate-controller"],"summary":"getCertificatesByTagName","operationId":"getCertificatesByTagNameUsingGET","produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/certificate/search/by-tags":{"get":{"tags":["gift-certificate-controller"],"summary":"searchForCertificatesBySeveralTags","operationId":"searchForCertificatesBySeveralTagsUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"},{"name":"tagsIDs","in":"query","description":"tagsIDs","required":true,"type":"array","items":{"type":"integer","format":"int64"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/certificate/sort":{"get":{"tags":["gift-certificate-controller"],"summary":"getSortedCertificates","operationId":"getSortedCertificatesUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"},{"name":"sort","in":"query","description":"Sorting criteria, e.g. -name,date. The '-' prefix indicates descending order. The default order is ascending. The first sorting criterion is name. Example: sort=-name,date","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/certificate/{id}":{"get":{"tags":["gift-certificate-controller"],"summary":"getById","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["gift-certificate-controller"],"summary":"deleteCertificate","operationId":"deleteCertificateUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["gift-certificate-controller"],"summary":"updateCertificate","operationId":"updateCertificateUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"giftCertificate","description":"giftCertificate","required":true,"schema":{"$ref":"#/definitions/GiftCertificate"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/order":{"get":{"tags":["order-controller"],"summary":"showAll","operationId":"showAllUsingGET_1","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/order/info/{id}":{"get":{"tags":["order-controller"],"summary":"getOrdersInfoByID","operationId":"getOrdersInfoByIDUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/tag":{"get":{"tags":["tag-controller"],"summary":"showAllTags","operationId":"showAllTagsUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["tag-controller"],"summary":"createTag","operationId":"createTagUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"tag","description":"tag","required":true,"schema":{"$ref":"#/definitions/Tag"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/tag/search/by-id/{id}":{"get":{"tags":["tag-controller"],"summary":"getById","operationId":"getByIdUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/tag/search/most-used":{"get":{"tags":["tag-controller"],"summary":"getMostUsed","operationId":"getMostUsedUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/tag/{id}":{"delete":{"tags":["tag-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/user":{"get":{"tags":["user-controller"],"summary":"showAll","operationId":"showAllUsingGET_2","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/by-user-id/{id}":{"get":{"tags":["user-controller"],"summary":"getOrdersByUsersID","operationId":"getOrdersByUsersIDUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"name":"page","in":"query","description":"page","required":false,"type":"integer","default":1,"format":"int32"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","default":3,"format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/create-order":{"post":{"tags":["user-controller"],"summary":"createNewOrder","operationId":"createNewOrderUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"order","description":"order","required":true,"schema":{"$ref":"#/definitions/Order"}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/user/search/{id}":{"get":{"tags":["user-controller"],"summary":"getUserById","operationId":"getUserByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"GiftCertificate":{"type":"object","properties":{"createDate":{"type":"string"},"description":{"type":"string"},"duration":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"lastUpdateDate":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"name":{"type":"string"},"price":{"type":"number"},"tags":{"type":"array","items":{"$ref":"#/definitions/Tag"}}},"title":"GiftCertificate"},"Link":{"type":"object","properties":{"deprecation":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"name":{"type":"string"},"profile":{"type":"string"},"rel":{"$ref":"#/definitions/LinkRelation"},"title":{"type":"string"},"type":{"type":"string"}},"title":"Link"},"LinkRelation":{"type":"object","title":"LinkRelation"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"Order":{"type":"object","properties":{"certificates":{"type":"array","items":{"$ref":"#/definitions/GiftCertificate"}},"closed":{"type":"boolean"},"cost":{"type":"number"},"createDate":{"type":"string"},"description":{"type":"string"},"id":{"type":"integer","format":"int64"},"lastUpdateDate":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"user":{"$ref":"#/definitions/User"}},"title":"Order"},"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"Tag"},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"links":{"type":"array","items":{"$ref":"#/definitions/Link"}},"name":{"type":"string"},"number":{"type":"string"},"orders":{"type":"array","items":{"$ref":"#/definitions/Order"}},"surname":{"type":"string"}},"title":"User"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}